GATEWAY-2947: query for document event description builder. - added request object builder. - moved base test to common lib - pom fixes for dependency management
