+++ Product/Production/pom.xml
@@ -57,6 +57,12 @@
       <artifactId>jmock-junit4</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+        <groupId>org.mockito</groupId>
+        <artifactId>mockito-all</artifactId>
+        <version>1.9.0</version>
+        <scope>test</scope>
+    </dependency>
   </dependencies>
   <build>
     <!-- Caution: If <resources> or <testResources> is overridden by
